chore: refactor internal circular dependencies
This change includes the nominal refactoring of the `dag-link` and `dag-node`
internal modules to remove the following circular dependencies:

 - `src/dag-link/index.js => src/dag-link/util.js => src/dag-link/index.js`
 - `src/dag-node/index.js => src/dag-node/create.js => src/util.js => src/dag-node/index.js`
 - `src/dag-node/index.js => src/dag-node/addLink.js => src/dag-node/index.js`
